+++ attributes/iostat_plugin.rb
@@ -1,0 +1,4 @@
+default['iostat']['path'] = '/usr/bin/iostat'
+default['iostat']['verbose'] = 'false'
+default['iostat']['python_folder'] = '/usr/share/collectd/python/'
+default['iostat']['include'] = '["tps", "kB_read/s", "kB_wrtn/s", "kB_read", "kB_wrtn", "rrqm/s", "wrqm/s", "r/s", "w/s", "rsec/s", "rkB/s", "wsec/s", "wkB/s", "avgrq-sz", "avgqu-sz", "await", "r_await", "w_await", "svctm", "%util"]'